requirements.txt
Flask==3.0.3
pyodbc==5.1.0
python-dotenv==1.0.1
requests==2.31.0